add i18n support (Spanish and Brazilian Portuguese)
languages are switched via a `locale=cc` parameter on any URL, where
`cc` is currently one of `en`, `es` or `pt`, with `en` as the default.

the method by which translated javascript files are loaded according to
the locale isn't the most elegant, yielding console errors if an unknown
locale is set, but at least the fallback file works correctly.